nfRendererD3D12: Refactor descriptor heap allocations
In D3D12 all descriptor allocations in heaps must be handled manually.
A helper class, HeapAllocator, has been introduced to help heap
management.

Verification: Build and run nfRendererDemo with nfRendererD3D12 backend.

+++ nfEngine/Renderers/nfRendererD3D12/ResourceBinding.hpp
@@ -49,8 +49,8 @@ class ResourceBindingInstance : public IResourceBindingInstance
 {
     friend class CommandBuffer;
 
-    size_t mDescriptorHeapOffset;
-    size_t mSamplerDescriptorHeapOffset;
+    uint32 mDescriptorHeapOffset;
+    uint32 mSamplerDescriptorHeapOffset;
     ResourceBindingSet* mSet;
 
 public:
